fix platform checks in the CLI code
Summary:
A couple places in the CLI code (mostly used by `eden doctor`) were checking
`sys.platform` to tell if we were on Linux.  Unfortunately these checks both
expected the value `linux2`.  However, since Python 3.3 `sys.platform` is just
`linux` on Linux, and not `linux2`.  This meant we were always hitting the
non-Linux code paths and skipping these checks.

This updates the code to check `platform.system()`.  Based on the
documentation it sounds like this is intended to give a bit more consistent
behavior across different platforms and OS versions.

Reviewed By: genevievehelsel

Differential Revision: D20178488

fbshipit-source-id: c908d5133a9c41e6a239a8893742d03f6c08527c