pyworker: ignore remove errors on Windows
Summary:
In some situations Windows may be unable to remove a file and Mercurial
cannot do anything about it. Instead of raising an error, just ignore it.

In the Python code, errors are ignored in multiple layers, but ultimately
batchremove ignore all OSError errors and just prints a warning about it. A
future diff will build a side channel for errors/warnings reporting, this will
be used to properly report this.

Reviewed By: quark-zju

Differential Revision: D20284803

fbshipit-source-id: 7df80f9e27e720ce819fb4116ea996fca390ad76